+++ app/client.py
@@ -12,7 +12,7 @@ from app.config import (
 
 import gspread
 from oauth2client.service_account import ServiceAccountCredentials
-from app.slack.models import Bookmark
+from app.models import Bookmark
 
 
 credentials = ServiceAccountCredentials.from_json_keyfile_dict(
